feat: add health check endpoint and update user model; enhance role handling and permissions

+++ app/routes/user_routes.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordRequestForm
-from starlette.authentication import UnauthenticatedUser
+from pymongo.errors import DuplicateKeyError
 
 from app.repository import user_repository
-from typing import List, Annotated
+from typing import List, Annotated, Optional
 from app.model.user import UserModel
 from app.routes.security import get_current_user, create_access_token, User, Token
 from app.dto.user.user_create_dto import UserCreate
@@ -22,8 +22,10 @@ async def list_users():
     except Exception as e:
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, str(e))
 
-@router.get("/{user_id}",)# response_model=UserModel)
-async def get_user(user_id: str):
+@router.get("/{user_id}", response_model=UserModel)
+async def get_user(user_id: str, current_user: Annotated[User, Depends(get_current_user)]):
+    if c.PERMISSION_READ_USER not in current_user.permissions and user_id != current_user.id:
+        raise HTTPException(status.HTTP_403_FORBIDDEN, "Insufficient permissions")
     try:
         user = user_repository.get_user_by_id(user_id)
         if user is None:
@@ -34,8 +36,9 @@ async def get_user(user_id: str):
 
 @router.post("", response_model=UserModel)
 async def create_user(
-        user: UserCreate,
-        current_user: Annotated[User | None, Depends(get_current_user)] = None):
+    user: UserCreate,
+    current_user: Annotated[User | None, Depends(get_current_user)] = None
+):# <--- Carinha triste ou brava?
     try:
         permissions = current_user.permissions if current_user else None
         return user_repository.create_user(user, permissions)
@@ -43,29 +46,33 @@ async def create_user(
         raise HTTPException(status.HTTP_400_BAD_REQUEST, str(e))
     except PermissionError as e:
         raise HTTPException(status.HTTP_403_FORBIDDEN, str(e))
+    except DuplicateKeyError as e:
+        raise HTTPException(status.HTTP_409_CONFLICT, "Duplicate email")
     except Exception as e:
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, str(e))
 
 @router.put("/{user_id}", response_model=UserModel)
-async def update_user(user_id: str, user: UserModel):
+async def update_user(user_id: str, user: UserModel, current_user: Annotated[User, Depends(get_current_user)]):
+    if c.PERMISSION_UPDATE_USER not in current_user.permissions:
+        raise HTTPException(status.HTTP_403_FORBIDDEN, "Insufficient permissions")
     try:
         return user_repository.update_user(user_id, user)
     except Exception as e:
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, str(e))
 
 @router.delete("/{user_id}", response_model=dict)
-async def delete_user(user_id: str):
+async def delete_user(user_id: str, current_user: Annotated[User, Depends(get_current_user)]):
+    if c.PERMISSION_DELETE_USER not in current_user.permissions:
+        raise HTTPException(status.HTTP_403_FORBIDDEN, "Insufficient permissions")
     try:
         return user_repository.delete_user(user_id)
     except Exception as e:
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, str(e))
 
-@router.get("/me/", response_model=User)
+@router.get("/me/", response_model=Optional[User])
 async def read_users_me(current_user: Annotated[User, Depends(get_current_user)]):
-    try:
-        return {"login": current_user}
-    except Exception as e:
-        raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, str(e))
+    #TODO: Remover depois
+    return current_user
 
 @router.post("/login", response_model=Token)
 async def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()]):
@@ -78,7 +85,13 @@ async def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()]):
                 headers={"WWW-Authenticate": "Bearer"},
             )
         else:
-            token = create_access_token(data={"sub": user.login, "scope": " ".join(form_data.scopes)})
+            token = create_access_token(data={
+                "sub": user.email,
+                "user_id": user.id,
+                "scope": " ".join(form_data.scopes),
+                "permissions": user.role.permissions,
+                "role": {"id": user.role.id, "name": user.role.name}
+            })
             return token
     except Exception as e:
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR, str(e))
